% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindCircularMap.R
\name{bindCircularMap}
\alias{bindCircularMap}
\title{Create a circular fault grid}
\usage{
bindCircularMap(rectangularMap, dieWidth, dieHeight, alignmentX = 0,
  alignmentY = 0, waferRay, outValue = NA)
}
\arguments{
\item{rectangularMap}{a sqare matrix}

\item{dieWidth}{the width of a chip}

\item{dieHeight}{the height of a chip}

\item{alignmentX}{the x coordinate of the center of the (0,0) chip}

\item{alignmentY}{the y coordinate of the center of the (0,0) chip}

\item{waferRay}{the ray of the wafer}

\item{outValue}{the value assignad to the element of of the circle. Default is NA}
}
\value{
the modified map
}
\description{
Creates a circular grid from a square one by inserting the value outValue in
every pixel which is not whitin the circle of given ray. The width and the height
of the pixel may be specified. Every pixel is identified as a point placed in the
center of its spatial extension. ie: if a chip's dimensions are 2x1, it is represent
as a point in (1;0.5).
}

