% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareWaferGrid.R
\name{prepareWaferGrid}
\alias{prepareWaferGrid}
\title{Prepare the axes for the wafer}
\usage{
prepareWaferGrid(dieWidth, dieHeight, waferRay)
}
\arguments{
\item{dieWidth}{the width of a chip}

\item{dieHeight}{the height of a chip}

\item{waferRay}{the ray of the wafer}
}
\value{
a list of 2 elemets, containing the x and the y axes.
}
\description{
Prepare a grid which is assumed to represent the wafer: the width and the height
of a single chip may be specified. A chip is represent as a point placed in the
middle of its spatial extension, ie: if a chip's dimensions are 2x1, it is represent
as a poin in (1;0.5). This funcion returns the x and the y axes of the wafer, one point
is placed for every chip.
}

